+++ CHANGELOG.md
@@ -12,9 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [NDMF] Non-destructive avatar conversion with NDMF and `VQT Avatar Converter Settings` component.
 - [NDMF] `VQT Platform Target Settings` component to specify the target platform for `VQT Platform Component Remover` and `VQT Platform GameObject Remover` components.
 - [NDMF] Warning message for using unsupported texture formats.
+- [NDMF] `VQT Network ID Assigner` component to assign Network IDs to the avatar.
 
 ### Changed
 - [NDMF] Duplicate meshes when removing vertex colors. Original meshes keep vertex colors.
+- `VQT Avatar Converter Settings` now uses VRChat SDK feature to assign Network IDs to the avatar.
 
 ### Removed
 - [NDMF] `Build Target` parameter from `VQT Platform Component Remover` and `VQT Platform GameObject Remover` components.